// src/_services/api/model/createProdutoDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Gestão de Devolução e Produtividade
 * Documentação da API responsável por: gestão de produtividade, montagem e impressão de mapas de separação, sistema de devolução de estoque e contagem de inventário.
 * OpenAPI spec version: 1.0
 */

export interface CreateProdutoDto {
  /** Código DUM do produto */
  codDum: string;
  /** Código EAN do produto */
  codEan: string;
  /** SKU do produto */
  sku: string;
  /** Descrição do produto */
  descricao: string;
  /** Shelf do produto */
  shelf: number;
  /** Peso líquido da caixa */
  pesoLiquidoCaixa: string;
  /** Peso líquido da unidade */
  pesoLiquidoUnidade: string;
  /** Unidades por caixa */
  unPorCaixa: number;
  /** Caixas por pallet */
  caixaPorPallet: number;
  /** Segmento do produto */
  segmento: string;
  /** Empresa do produto */
  empresa: string;
  /** Tipo de peso */
  tipoPeso: string;
  /** Criado em */
  criadoEm: string;
}